+++ packages/design-system/src/components/DateField/index.ts
@@ -1,2 +1,2 @@
-export { default as DateField } from './DateField';
+export { default as DateField } from './MultiInputDateField';
 export { default as DateInput } from './DateInput';
